requirements.txt
Django==3.0.3
djangorestframework==3.11.0
django-cors-headers==3.2.1
djangorestframework-jwt==1.11.0
django-webpack-loader==0.7.0
channels==2.4.0